backend/requirements.txt
fastapi
uvicorn
pydantic
langchain
langchain-google-genai
google-ai-generativelanguage
python-dotenv
langchain-community